requirements.txt
Flask
flask-restplus
git+https://github.com/simeon-xx/flask-restplus-patched.git
marshmallow
flask-mongoengine
marshmallow-mongoengine
webargs
